Revert "refactor: split up cli into separate modules"
This reverts commit 4638aeccecd145b5ff5af99d152778d7de2c079b.
