+++ app/src/cli/new_project.rs
@@ -5,17 +5,11 @@ use ambient_project::SnakeCaseIdentifier;
 use anyhow::Context;
 use convert_case::Casing;
 
-use super::ProjectPath;
-
-pub(crate) fn handle(
-    project_path: &ProjectPath,
+pub(crate) fn new_project(
+    project_path: &Path,
     name: Option<&str>,
     api_path: Option<&str>,
 ) -> anyhow::Result<()> {
-    let Some(project_path) = &project_path.fs_path else {
-        anyhow::bail!("Cannot create project in a remote directory.");
-    };
-
     // Build the identifier.
     let project_path = if let Some(name) = name {
         project_path.join(name)
